% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.R
\name{backtr}
\alias{backtr}
\title{Back transformation function}
\usage{
backtr(scores, nscore, tails = "none", draw = TRUE)
}
\arguments{
\item{scores}{a vector of normal scores and a normal score object (from nscore)}

\item{nscore}{the nscore object}

\item{tails}{One major issue is how to extrapolate to the tails. Options}

\item{draw}{show qq plot}
}
\value{
numeric
}
\description{
Given a vector of normal scores and a normal score object
}
