% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.R
\name{modelSoftHard}
\alias{modelSoftHard}
\title{Builds stats}
\usage{
modelSoftHard(hard, soft)
}
\arguments{
\item{hard}{numeric vector}

\item{soft}{numeric vector}
}
\value{
model object
}
\description{
Given a hard (primary) dataset and a soft (secondary) vector
}
