% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.R
\name{write.data.frame.geoeas}
\alias{write.data.frame.geoeas}
\title{Writes a data.frame to a GEO_EAS text file}
\usage{
write.data.frame.geoeas(dat, outfile)
}
\arguments{
\item{dat}{data.frame}

\item{outfile}{character, file path}
}
\description{
This function writes a data frame to a GEO-EAS text file.
Useful for interacting with GSLIB.
}
